Fix hipsolver-test on Windows (#42) (#43)
* Copy hipsolver.dll into clients/staging directory

* Undo removal of CONFIG from find_package(hip)

This reverts a change in e87ec62cf67ab8534782af879f9604ec7f86d11b that
should not have been made. The existence of both hip-config.cmake and
FindHIP.cmake makes this code unusual, and it should remain as it was,
to match hipBLAS and other hip marshaling libraries. Though, it has no
effect on the build either way, as the use of PATH implies CONFIG.

[ROCm/hipSOLVER commit: 344ef85f72c42214788ab21653c11e5eef3f0a6d]