Fix docs parameter formatting (#318)
* Add ReadTheDocs config file

The ReadTheDocs developers recommend using a configuration file.
When .readthedocs.yaml is present, the 'Advanced settings -> Default settings'
section of the ReadTheDocs admin panel is ignored and the configuration
file is used instead.

The yaml file has some benefits when compared to the web-based
settings panel:

- If you move conf.py or requirements.txt, you can update
  .readthedocs.yaml at the same time. That way, you retain the ability 
  to easily rebuild old versions of the docs, because ReadTheDocs will
  automatically use the corresponding settings.

- External contributors can see how we have configured ReadTheDocs,
  which may help them in understanding our project structure with
  respect to documentation.

- There are some settings you can change via the yaml file that you
  cannot change via the web interface.

* Specify version requirements more precisely

The docs for 4.1.0 were rebuilt a few months after release, and they
changed in appearance (for the worse). Examining the build logs, it
seems that the packages used for the build changed from Sphinx-3.5.3
and breathe-4.28.0 to Sphinx-4.0.2 and breathe-4.30.0.

Following the ReadTheDocs advice for Reproducable Builds [1], I'm now
specifying exact versions for our key dependencies.

[1]: https://docs.readthedocs.io/en/stable/guides/reproducible-builds.html

* Turn sphinx warnings into errors

* Specify docutils 0.16

There are conflicts between docutils >=0.17 and our sphinx style,
so use the last working version. This fixes the rendering of lists
in the rocsolver documentation.

[ROCm/rocSOLVER commit: 0777bc012e2b853e13bbe721a746568651ba2530]