Minor client platform fixes (#1072)
* Replace sys_siglist with sigdescr_np

glibc 2.32 removed sys_siglist and provided sigdescr_np as a
replacement.

* Remove conflicting namespace declaration

It's not valid to inject a definition into the std namespace. In later
versions of libstdc++, the <chrono> header adds symbols into the
std::filesystem namespace. As such, std::filesystem cannot then be
redefined to be an alias of the std::experimental::filesystem namespace.

Fixes the error:

    /root/rocBLAS-rocm-4.5.2/clients/common/utility.cpp:36:15: error: redefinition of 'filesystem'
        namespace filesystem = experimental::filesystem;
                  ^
    /usr/bin/../lib/gcc/x86_64-linux-gnu/11/../../../../include/c++/11/chrono:3343:13: note: previous definition is here
      namespace filesystem
                ^

* Use python3 from PATH in rocblas_gentest.py

Using #!/usr/bin/python3 as the shebang hardcodes gentest to the
system Python. Searching for Python via env provides more flexibility.

[ROCm/rocblas commit: 3738f8b098cdc1db1bdfc164ceb689d073116c98]