Move gfx11 to the optional targets (#480)
This changes rocSOLVER to only add gfx11 to the default build targets
when rocm_check_target_ids reports that they are supported by the
compiler. It enables rocSOLVER to be built more easily on older
compilers. Additionally, the check for compiler support is skipped
when an AMDGPU_TARGET has been explicitly specified (since its only
purpose is for setting the default value, which is not used in that
case).

[ROCm/rocSOLVER commit: d83f624c6145f38d693189be64667ae928d60936]