There may be a case where the local UID collides with a docker UID for the non-root user. As long as the UID:GID is the same as the directory and file permissions this should be totally fine. This can happen with VSCode from the base docker image adding its user mirroring the $USER UID of the local user.
The exist behavior is for such a collision to cause a the `user-entrypoint` script to fail when adding a user via `useradd`.
This PR adds the flag `-o` to the `useradd` command such that the command will not fail in the above case and allows the docker container to run.

Co-authored-by: dlowell <dlowell@amd.com>

[ROCm/rocRoller commit: c4bd8301d668fadf66141ce5575317e624dec1f5]