Fix missing sqlite include paths (#1001)
rtc_cache.h includes sqlite3.h, so anything that uses rocfft-rtc-cache
needs to have the sqlite3 include directories. That is typically best
accomplished with a target_link_libraries declaration, as the include
path for the library will automatically be propagated to the dependent
library.

In modern CMake, it's very rare to have to add the include directories
of a library that your code is using. For any properly specified
library, you should be able to use target_link_libraries and the
required include paths for the library will be added to the target's
build options.

[ROCm/rocFFT commit: f66855d5a337f3aa4e9be44fd639855fd301c9c0]