Fix sparse test failures on Windows (#183)
The Windows CI only copies <builddir>/clients/staging to the test
machine, so the data files must be copied to that location. The
search for the sparse test data is also restricted to exe-relative
lookups, matching the behaviour in rocSOLVER as of commit 1361711.

(cherry picked from commit 5425826b43b3fe8413356fb8c676229201106db3)


[ROCm/hipSOLVER commit: 50df6c17f55b6cb70e0aedbb90c18a08cf069e78]