Split AddLDS transform into: an early AddLDS transform and a late AddPrefetch transform.
The AddLDS transform simply changes, eg, LoadTiled into LoadTiled + StoreLDSTile + LoadLDSTile.  This is done early.  Doing this means that the coordinate transforms added in `LowerTile` are easier to reason about.  Unfortunately, this means that LowerTensorContraction becomes more complicated, as it now has to reason about StoreLDSTile and LoadLDSTile.

The "deduplication" step of FuseLoops has been moved into it's own transform: RemoveDuplicates.

A new helper `colourByUnroll` has been added.  This is used to do graph colouring; and it is now used in RemoveDuplicates, OrderEplilogueBlocks, and the updated AddPrefetch.

[ROCm/rocRoller commit: 1a069388208a3770901ca1216db3b0269dba36e2]