Plug in manager (#43)
* add plug in manager intro

* format code

* made map to vector

* made map to vector

* address review comments

* review comments

* more review comments

* add plugin to cmake

* add execute plan

* Forward declare the Engine_config_descriptor class

And update a few comments.

* Change plugin finalize() to get_max_workspace_size()

The engine has already been finalized at this point.

Instead, the plugin manager will need to select the correct plugin
based on the global index set on the engine descriptor. The plugin
manager then needs to query this workspace size and set that size on
the engine config descriptor.

* Add HIPDNN_ENGINE_ID_ prefix to the engine IDs

* Remove engine and engine config descriptors for now

These can be fleshed out later.

* Add graph desc and engine id arguments to get_max_workspace_size

* add review comments

* typo

* more comments

* format fix

---------

Co-authored-by: Bibek Ghimire <bghimire@amd.com>
Co-authored-by: Jonathan Lichtner <195780826+JonathanLichtnerAMD@users.noreply.github.com>

[ROCm/hipDNN commit: bbd562e0b24ce8874391f40ba21fe9f3cdb5c8ce]