[Ex CI] always create an Azure summary check (#691)
Currently the summary check will not be created if no runs were
triggered, but the check needs to exist if it is to be a blocking
requirement. Adds a snippet in the dispatcher to create a 'neutral'
summary check if no runs were triggered.

Also disables the dispatcher from running on draft PRs. It will still
run whenever a draft PR is converted to ready for review.

Test PR: https://github.com/ROCm/rocm-libraries/pull/690
Sample neutral check:
https://github.com/ROCm/rocm-libraries/pull/690/checks?check_run_id=46047987130