Changing 128-bit assert_eq function to macro (#965)
The issue with assert_eq() as the function it is written now is that it
cannot break the validation for loop if it detects a failure, resulting
in potentially unlimited failure outputs to console. This is because any
break called inside assert_eq is limited to the scope of the function.

I have changed assert_eq() to the ASSERT_EQ_128BIT_WORKAROUND macro.
This preserves the scoping, and thus when gtest's ASSERT_EQ fails and
triggers a break, it will break the for loop.

I also took out the "U" datatype check, as I do not currently see a
scenario in the block_radix_sort unit tests where the output datatype
and the expected datatype differ.