[hipblas] rewrite code coverage logic to use pure LLVM (#1145)
CI was failing in monorepo. Noticed that we are trying to use gcov and
llvm-cov together, so I simplified the code to make it only use llvm-cov
and did some refactoring to decouple the CMake code from Jenkins.

Also removed `construct_build_prefix` from .jenkins folder because it is
not needed. We already do that in rocJenkins before the project is
loaded.

Code cov CI is passing with these changes:
http://math-ci.amd.com/job/rocm-libraries/job/codecov/job/hipblas/view/change-requests/job/PR-1145/

Here is the coverage report:
http://math-ci.amd.com/job/rocm-libraries/job/codecov/job/hipblas/view/change-requests/job/PR-1145/Code_20coverage_20report/

And screenshot for quick reference:
<img width="645" height="328" alt="image"
src="https://github.com/user-attachments/assets/2f9455ab-26d4-45c6-9959-c1ce892d408e"
/>