Performance optimization for histogram with large bin count on mi300x (#719)
# Context
A performance optimization for bins that do not fit in shared memory on
mi300x. It also includes some cleaning up of the code and some bug fixes
for the benchmarks. The graph also includes the other cases where it
does fit in shared memory.

It also includes a commit to make the human output better readable,
https://github.com/ROCm/rocm-libraries/pull/719/commits/1cbdf1d23844ff0e038712d7aadd34b4dc4fb1f1.

Results:

![graph_device_histogram_rocprim_mi300x](https://github.com/user-attachments/assets/99c78b4c-2550-4946-b57b-5c4c44f8bdbe)