rocThrust CCCL 2.8.5 changes (#1100)
Most of the changes here are pulling in the commits from CCCL 2.8.5.
With the fixes and workarounds attached to these commits.

In addition, this PR also includes:
- Introduction of an opt-in mechanism to treat libhipcxx as a soft
dependency (temporarily disabled the path using libhipcxx)
- Clean-up of thrust/detail/type_traits.h
- Addition of missing rocThrust tests

**_PLEASE DO NOT SQUASH THE COMMITS IN THIS PR_**, as CCCL 3.0.0 will
branch off from the middle of this series and the commit history needs
to be preserved.

---------

Co-authored-by: Nara Prasetya <nara@streamhpc.com>