Add new MFMACoexecObserver and simpler cost weights. (#1294)
## Motivation

- Add a new MFMACoexecObserver which is aware of other instruction types
and which ones can execute during the cycles following an MFMA.
- So far this is only configured for `v_mfma_scale_f32_16x16x128_f8f6f4`
and only for the non-8-bit types.
- This required the addition of the `CoexecCategory` enum which can be
gotten from an opcode through the `GPUInstructionInfo` class or from the
`getCategory()` accessor of an `Instruction`.
- Enable a much simpler set of scheduler cost weights, which can be set
by setting the cost function to `LinearWeightedSimple`.

## Details

I've now clarified the presence of the `nothing` cycle in for the
`v_mfma_scale_` case: It is to make sure we count these as two
instructions for the purposes of this and other observers' bookkeeping.
It's only needed for the `mfma_scale` instructions, not plain `mfma`
instructions.