[hipblaslt] Completed and enabled swizzleB feature
## Motivation

<!-- Explain the purpose of this PR and the goals it aims to achieve.
-->
SWDEV-509997: 
SwizzleB was implemented about 95%, it was planned in 6.5 but since no
urgent request so we haven't finished the last 5%.
The rest of features about swizzle-b is LSU, hipblaslt host-side
pipeline, and a few basic logic yamls.
Since this feature is almost implemented, I'd like to make it fully
completed so that we can make some progress for the ticket.

## Technical Details

<!-- Explain the changes along with any relevant GitHub links. -->
Implemented swizzle-B, aligned with swizzle-A.
~~Updated: LSU would be implemented in another PR.~~

## Test Plan

<!-- Explain any relevant testing done to verify this PR. -->
tox will test swizzleB.yaml, and will put a basic logic yaml in gtest.

## Test Result

<!-- Briefly summarize test outcomes. -->

## Submission Checklist

- [x] Look over the contributing guidelines at
https://github.com/ROCm/ROCm/blob/develop/CONTRIBUTING.md#pull-requests.
- [x] Swizzle-B: VW
- [x] Swizzle-B: Data-Types
- [x] Swizzle-B: LSU
- [x] Swizzle-B: Passed tox tests
- [x] Swizzle-B: Supported hipblaslt host side 
- [x] Swizzle-B: Passed gtest