[hipblaslt] hipblaslt version and find blis patch (#1596)
## Motivation

The versioning functionality provided by rocm-cmake is inflexible. This
PR allows for a custom version information to be set at configure time.
If not, it falls back on the rocm-cmake approach. Further, the find blis
module does not account for flat build directories without the extra
`CMAKE_BUILD_TYPE` directory. this adds a path to enable finding blis
when installed into the `build/deps` directory in the new build system.

## Technical Details

- Adds cache variable HIPBLASLT_PROJECT_VERSION and sets default value
to 1.1.0.
- When HIPBLASLT_PROJECT_VERSION is passed at configure time set major,
minor and patch using first there entries.
- Checks if there is a fourth entry for tweak version and if not set
uses rocm-cmake to compute a shortened hash for tweak.
- Adds build/deps to paths in findblis.cmake such that the package is
discoverable when using cmake directly rather than the install script.

## Test Plan

Pipelines and numerous local tests.