fix(hipcub): fix incorrect args in cmake (#1719)
## Motivation

This fixes two issues relating to the build configuration in hipCUB:
1. Offload compression is broken due to `amdclang` no longer globally
accepting HIP-related flags.
2. Fixes interaction between non-HIP environments and CMake. This also
fixes issues where the host compiler is set to a non-standard CXX
compiler (i.e. not GCC nor Clang). In that case we also emit a message
as this was in a non-working state before this change as well.

## Technical Details

The first fix is quite simple and is resolved by passing `-xhip` to the
check to ensure it is checked as it were a HIP-file.

For the second fix we'll leverage the fact we can use CMake's
`enable_language(...)` due to our set minimum CMake version.

## Test Plan

Configure and build pass.

## Test Result

Works on my machine (tm), except for a few cases that are fixed by
current open PRs.

## Submission Checklist

- [x] Look over the contributing guidelines at
https://github.com/ROCm/ROCm/blob/develop/CONTRIBUTING.md#pull-requests.