[hipblaslt] Avoid installing rocm-cmake if already found (#1627)
## Motivation

It seems that `hipblaslt` always download rocm-cmake, even if it is
already available in the system or environment.

## Technical Details

At first glance, it seems that this was only due to a type, in which
`ROCM_FOUND` was accidentally used instead of
`ROCmCMakeBuildTools_FOUND`.

## Test Plan

I built the project, and verified that if `rocm-cmake` was already
installed, it was not downloaded again.

## Test Result

See test plan.

## Submission Checklist

- [x] Look over the contributing guidelines at
https://github.com/ROCm/ROCm/blob/develop/CONTRIBUTING.md#pull-requests.