[rocrand] Add a new CMake option to allow tests to be built from installed ROCm libraries (#1852)
## Motivation

<!-- Explain the purpose of this PR and the goals it aims to achieve.
-->
There is a request to be able to build `rocrand` tests from installed
ROCm, rather than from the source tree. This PR addresses the request.

## Technical Details

<!-- Explain the changes along with any relevant GitHub links. -->
A new CMake option, `USE_SYSTEM_LIB` is introduced to support this
request. This option will take effect when the `BUILD_TEST` option is
set to `ON`. By default, it is `OFF`.

## Test Plan

<!-- Explain any relevant testing done to verify this PR. -->
Enabling `USE_SYSTEM_LIB` will change the include and link paths that
the compiler searches for the `rocrand` library.

## Test Result

<!-- Briefly summarize test outcomes. -->
The compile and link commands were verified that the `rocrand` library
was picked up from the correct location.

## Submission Checklist

- [ ] Look over the contributing guidelines at
https://github.com/ROCm/ROCm/blob/develop/CONTRIBUTING.md#pull-requests.