Set origami install to off (#2179)
## Motivation

The goal was to prevent the origami license folder from being installed.

## Technical Details

Added `set(ORIGAMI_ENABLE_INSTALL OFF)` to the top-level CMakeLists.txt
file before the origami subdirectory is added. This is a **temporary
solution** since the origami build system does not currently provide a
way to install the library without also installing the documentation and
license files.

Previously PR #2078.

## Test Plan

N/A

## Test Result

N/A

## Submission Checklist

- [ ] Look over the contributing guidelines at
https://github.com/ROCm/ROCm/blob/develop/CONTRIBUTING.md#pull-requests.