[HipDNN]: Unify testing seeds (#2205)
Motivation
With adding hipDNN to TheRock we are seeing infrequent tolerance issues
in the integration tests due to randomized seeds.
This PR locks the random seeds to ensure tests are stable.
This matches with how MIOpen is currently testing as well.

Technical Details
Since we are validating at an integration level, we cannot validate too
strictly when comparing the results to a reference implementation. There
is various factors that impact the final results (atomics leading to
non-deterministic behavior, algorithm choices etc). The integration
tests are meant to give us confidence that the integration is still
working and not broken, but it cannot provide very strict tolerance
ranges due to this variability, and lack of detail.

We plan to improve on this in the future, and provide more details about
the selected implementation to allow more tight tolerance checks.

Test Plan
Run existing tests on multiple architectures to ensure they are
validating.
Run tests with seeds and logging to ensure it's using environment
variable.
Run tests for new file to ensure it's working as expected.
Run tests with new env variable set to ensure it's not sensitive or
polluting the env.

Test Result
All tests passing.