Runtime Configurable WGMXCC (#2170)
## Motivation

This PR enables runtime configurable WGMXCC so libraries can be shared
across different architectures without changing the current state of
libraries.

## Technical Details

Prior to this PR, there were two versions of WGMXCC (StreamKXCCMapping
and WorkGroupMappingXCC), it was possible to have both present in one
kernel. Although both were supposed to the same thing, there was slight
differences between the two in terms of the implementation. For example,
WorkGroupMappingXCC checks to have the same WGs per group and avoids
mapping if the remaining WGs cannot be distributed evenly in the groups.
Moreover, it also distributes according to the number of CUs in some
cases.
In this PR, `WorkGroupMappingXCC: -1` is introduced. This parameter uses
the same algorithm as in `StreamKXCCMapping`, but the value can be
configured at runtime to be equal to the number of XCCs on the
architecture. We can also use Origami prediction to change this value at
runtime. Moreover, an env variable is introduced to change the runtime
WGMXCC to any value in the acceptable range for this parameter.

## Test Plan

Tested locally and CI.

## Test Result

-

## Submission Checklist

- [ ] Look over the contributing guidelines at
https://github.com/ROCm/ROCm/blob/develop/CONTRIBUTING.md#pull-requests.