Add deterministic tie breaker logic for Origami (#2264)
## Motivation
Currently the tie-breaker logic in origami is non-deterministic with
respect to the permutation (ordering) of the input solution list.

## Technical Details
To make this tie-breaking deterministic I have added logic which
performs an additional tie break (after the original one) on problem
dimensions, and then just consistently selects the tile with the largest
M (will only apply in square problems).

## Test Plan
Added an origami test which permutes the inputs and expects the same
output (failed before, doesn't fail now). We also need to make sure the
perf is not significantly worse for some reason using OOB benchmarks.

## Test Result
The test passes now, OOB benchmarks pending.