Add Stream K experimental support for RDNA (#1554)
This PR introduces initial changes needed for stream-k RDNA support.
Some experimental tuning was conducted for a specific model, and it is
working. Now we are conducting additional tuning and doing performance
analysis.

Major changes:
- The s_store instruction needed to be swapped with buffer_store since
s_store is not supported on RDNA
- Additional modifiers were added to rocisa that were missing for gfx11
and gfx12 and an "off" option was added for vgpr
- Add gfx1100 and gfx1201 architectures to Origami Analytics