[rocRoller] Add more simplify passes to FastArithmetic pipeline (#2336)
## Motivation

The `simplify` expression transformation is currently called twice in
the `FastArithmetic` pipeline, once explicitly and once inside
`splitBitfieldCombine`. This PR adds another simplify call at the end of
the pipeline, as it further simplifies expressions present in GEMM
kernels.

It seems that the generated code does not change because the
`FastArithmetic` is called twice by the code generator, but it may
benefit other places in the code that call `FastArithmetic` only once.

## Technical Details

- Adds a simplify call to `FastArithmetic` 
- Adds `applyTransforms` and `getTransforms` functions to
`FastArithmetic` to enable testing with different pipelines

## Test Plan

- Adds test "FastArithmetic pipeline properly simplifies expressions"
with an example that benefits from the extra simplify