[hipDNN] Fix tensor Ids being filtered if they were 0 (#2462)
## Motivation

Pointwise binary operations were incorrectly filtering tensor Uids that
were 0.
This could lead to graph failures for testing and samples, depending on
how the UIds are assigned.

## Technical Details

std::unordered_set was being used during tensorID population, and if
specifically the optional TensorId was being set to 0, it wouldn't get
propagated properly. This lead to flaky breaks in tests and samples if
the Uids were allowed to be set automatically.

## Test Plan

Build, run all the tests, and ensure they pass.

## Test Result

Tests passed!