[rocRoller] Use rocprofiler-sdk for testing cycle count prediction (#2052)
## Motivation

Integrate
[rocprofiler-sdk](https://rocm.docs.amd.com/projects/rocprofiler-sdk/en/latest/index.html)
into test framework. Enables future validation of observers and other
predictor tools.

<!-- Explain the purpose of this PR and the goals it aims to achieve.
-->

## Technical Details

- The tests are only built and included if
`-DROCROLLER_ENABLE_ROCPROF_ANALYZER=ON` is passed to CMake
- New dependencies added to the dockerfiles
- A per-process agent is added to interact with the rocprofiler APIs

<!-- Explain the changes along with any relevant GitHub links. -->

## Test Plan

- Added tests to ensure profile for correct kernel is returned, due to
the global nature of the agent
- Per discussion, in the future, these tests will be included as a part
of a "risky tests" CI job -- the concern is that these tests can fail to
build if the rocprofiler-sdk interface changes due to a new amdgpu build

<!-- Explain any relevant testing done to verify this PR. -->

## Test Result

Tests work as expected. With `-DROCROLLER_ENABLE_ROCPROF_ANALYZER=ON`:

`tests/rocroller-tests-rocprofiler` / `ctest -R .*Rocprofiler.*`

<!-- Briefly summarize test outcomes. -->