[rocRoller] Enable using Raw32 types in SplitBitfieldCombine transformation (#2839)
The SplitBitfieldCombine transformation was extracting dwords from
Expressions as UInt32, this PRs fixes that so it uses Raw32.
As a result, other simplifications that would create `Convert`
expressions to Raw32 stop working as Conversion expressions do not
support the Raw32 type.

The rest of the PR enables temporarily disables simplifications that
would add conversions to Raw32 at the expression level, does some of
these conversions during code generation, and adds a reinterpret
function that replaces translation-time conversion to Raw32.