[MIOpen] Porting Neuron kernels from OpenCL to HIP (#1908)
## Motivation

This PR is part of the effort to port remaining OpenCL kernels in MIOpen
to HIP. This pull request focuses on porting Neuron activation kernels.

## Technical Details

- The following kernels are ported to HIP:
    - MIOpenActiveBwdLite
    - MIOpenActiveBwd2DLite
    - MIOpenNeuronBwd
    - MIOpenActiveFwdLite
    - MIOpenActiveFwd2DLite
    - MIOpenNeuronFwd
- Removed excessive macro usage, replaced with if constexpr.
- Refactored to avoid code duplication.
- Added a `miopen_math.hpp` header where math functions are overloaded
for different types.
- FP16 tests are added to `gtest_activation`.
- Increased elements per thread when FP16.

## Test Plan

Changes can be verified with `test_gtest_activation` and
`test_unit_ActivationDescriptor`. The non-lite kernels need to be tested
with `MIOPEN_DEBUG_FIND_ONLY_SOLVER="ActivFwdSolver1;ActivBwdSolver1"`
enabled and returning true in the IsApplicable() functions of the
solver.

## Test Result

Tests pass on gfx90a

## Submission Checklist

- [x] Look over the contributing guidelines at
https://github.com/ROCm/ROCm/blob/develop/CONTRIBUTING.md#pull-requests.

---------

Co-authored-by: SreecharanGundaboluAMD <sgundabo@amd.com>