[MIOpen] Port CTCLoss Kernel from OpenCL to HIP (#2196)
## Motivation

This PR is part of the effort to port remaining OpenCL kernels in MIOpen
to HIP, in particular porting the CTC Loss kernel.

## Technical Details

* Ports `MIOpenCTCLoss` to HIP.
* Removes OpenCL host code from `ctcolc.cpp`.This includes the only path
where the `OPT_LCL_MEM_GRAD` macro could be defined for kernel code, and
the define and preprocessor guarded code path was therefore removed from
the kernel code as part of the port.
* Updated data types to use `FLOAT` definition from the `float_types.h`
* Removed type punning from `AtomicLogAddExp` and replaced with device
`memcpy` calls to unsigned type of same bit width.
* Type casts to pointer types that are no longer required have been
removed.

## Test Plan

Changes were verified with `test_ctc` and `MIOpenDriver ctc -V 1`, which
cover the FP32 data type only.

## Test Result

Tests pass on gfx906, gfx908, and gfx90a.

## Submission Checklist

- [x] Look over the contributing guidelines at
https://github.com/ROCm/ROCm/blob/develop/CONTRIBUTING.md#pull-requests.