Fix conversion from rocisa DataType to origami data_type_t (#3125)
## Motivation

Fix a bug converting rocisa DataType to origami data_type_t.

## Technical Details

A static_cast is not sufficient, since the types don't map 1:1. Move the
conversion function to a utility file and use for all instances where a
conversion is needed. This fixes a bug where Origami would use the wrong
latency values for selection of some types.

## Test Plan

Origami was printing warning messages of invalid data types. Verified
this patch fixes the issue.