Port fusion_aux to GTest (#2398)
## Motivation

This PR converts the fusion_aux test from CTest to GTest.

## Technical Details

The test itself is unchanged, but the source code has been modified to
use the GTest framework.

## Test Plan

Execute the new GTest version and verify it is successful like the
previous one.

## Test Result

All the test have passed. The test results are attached to this PR.

[test_fusion_aux.txt](https://github.com/user-attachments/files/23265847/test_fusion_aux.txt)

## Submission Checklist

- [X] Look over the contributing guidelines at
https://github.com/ROCm/ROCm/blob/develop/CONTRIBUTING.md#pull-requests.

---------

Co-authored-by: JonathanLichtnerAMD <195780826+JonathanLichtnerAMD@users.noreply.github.com>