[hipDNN] [ALMIOPEN-406] Add applicability checks to fused CBA and CA kernels (#3186)
## Motivation

Many of the applicability checks are handled by MIOpen, but some of
these checks require knowledge of the hipdnn graph and must be handled
directly in the plugin. In addition to adding these checks, this PR adds
testing to cover a wide range of graphs which should be accepted and
rejected.

## Technical Details

Changes:
- Added isApplicable checks for: compute data type and intermediate data
types, (edited to add) **and activation checks**
- The `ConvFwdBiasActivation` tests have been consolidated into a single
suite with the intent of making it easy to expand in the future.
- `hipdnn_frontend` has been added as a dependency for the miopen plugin
tests
- Added MockNode
- Moved global applicability check for a float compute type into
individual nodes since CBA allows for non-float compute types in bias
- Modified `test_name_validator.py` to distinguish ellipses from the
period after a test suite name

## TO DO

~~Once #3047 is merged, add to the documentation and update sample to
use correct compute data type for bias~~
Done!

## Test Result

Tests and samples all pass locally

## Submission Checklist

- [ ] Look over the contributing guidelines at
https://github.com/ROCm/ROCm/blob/develop/CONTRIBUTING.md#pull-requests.