[hipBLASLt] [rocRoller] Create a SolutionCache class to replace nested map data structure (#2732)
## Motivation

This PR should not change behavior of selected kernel. It is cleaning up
implementation by creating a SolutionCache. This way internal cache can
change in the future without the rest of the code needing to be changed.

## Technical Details

Cache is currently using nested maps. There are now methods for adding a
solution to the cache and retrieving a solution.

Solution Cache is using unordered_maps. In order to use unordered_maps,
it was necesary to add hash functions for the KernelType class. To make
this easier, a ScaleType struct was added as well.

## Test Plan

Run current tests, behavior should not change.

## Submission Checklist

- [x] Look over the contributing guidelines at
https://github.com/ROCm/ROCm/blob/develop/CONTRIBUTING.md#pull-requests.