[MIOpen] Port bn_peract_test to GTest (#2503)
## Motivation

This PR converts the bn_peract_test test from CTest to GTest.

## Technical Details

The test itself is unchanged, but the source code has been modified to
use the GTest framework.

## Test Plan

Execute the new GTest version and verify it is successful like the
previous one.

## Test Result

All the test have passed. The test results are attached to this PR.

[test_bn_peract_test.txt](https://github.com/user-attachments/files/23397002/test_bn_peract_test.txt)

## Submission Checklist

- [X] Look over the contributing guidelines at
https://github.com/ROCm/ROCm/blob/develop/CONTRIBUTING.md#pull-requests.

---------

Co-authored-by: SreecharanGundaboluAMD <sgundabo@amd.com>