[hipDNN] Add CMake presets and improve toolchain detection. (#3210)
## Motivation

CMake presets can simplify the effort needed to manage the
configuration, building, and testing phases of development.

Also want to improve the flexibility of the toolchain automatic
detection on windows computers.

## Technical Details

Two preset files are added, one to projects/hipdnn, and one to
projects/hipdnn/samples. Use cmake --list-presets to view the pre-made
presets. Run ctest --list-presets to view the pre-made test presets.

Additional CMake variables to assist in auto-discovering the required
tools:
* `LLVM_TOOLS_SEARCH_PREFIX` - set to the directory prefix that the tool
version will be appended to. I.e. `c:\dist\clang` will be updated to
include folders `c:\dist\clang18` (for clang-format) and
`c:\dist\clang20` (for clang-tidy and others).
* `ROCM_CMAKE_PATH` - set to the ROCm/TheRock install folder (not
including `\bin`). Can also be set in the environment instead of passed
with -D.
* `ROCM_CMAKE_HIPCOFING_PATH` - the folder that hipconfig.exe is located
in. In this case hipconfig.exe will be used to locate the ROCm toolchain
(ROCm does not need to be included in the system path).

On Windows the ROCm bin folder will need to be in the system path so
that programs can find the needed dynamic libraries. With the bin folder
in the path, the `LLVM_TOOLS_SEARCH_PREFIX` can be used to set the
location of the additional Clang-based tools. This would typically be
the minimal set-up for building hipDNN on Windows.