[hipDNN]: Add batchnorm inference variance schema (#3494)
## Motivation
We need to support batchnorm with variance instead of inv_variance. The
best way to do this is to have completely separate schemas for them so
its not confusing to the end user which one they are using.

## Technical Details

This just adds the schema so the rest of the work can be split up
easier.

## Test Plan

<!-- Explain any relevant testing done to verify this PR. -->

## Test Result

<!-- Briefly summarize test outcomes. -->

## Submission Checklist

- [ ] Look over the contributing guidelines at
https://github.com/ROCm/ROCm/blob/develop/CONTRIBUTING.md#pull-requests.